backend/requirements.txt
SQLAlchemy~=2.0.34
fastapi~=0.114.0
pydantic~=2.9.1
pydantic-settings~=2.5.2
uvicorn
jieba
requests
redis
alembic
passlib[bcrypt]
pyjwt